refactor(walkthrough): rename and move the schema
